Don't empty bucket if no objects

+++ dev-tools/src/lambdas/emptyS3Buckets/emptyS3Bucket.ts
@@ -8,18 +8,23 @@ export const emptyS3Bucket = async (bucketName: string): Promise<void> => {
     `Found ${objects.versions.length + objects.deleteMarkers.length} objects`,
     { bucketName }
   )
-  await Promise.all(
-    objects.deleteMarkers.map((object) =>
-      deleteObject(bucketName, object.Key, object.VersionId)
-    )
-  )
-  await Promise.all(
-    objects.versions.map((object) =>
-      deleteObject(bucketName, object.Key, object.VersionId)
-    )
-  )
 
-  logger.info('Successfully emptied S3 bucket', { bucketName })
+  if (objects.versions.length === 0 && objects.deleteMarkers.length === 0) {
+    logger.info('No objects to delete', { bucketName })
+  } else {
+    await Promise.all(
+      objects.deleteMarkers.map((object) =>
+        deleteObject(bucketName, object.Key, object.VersionId)
+      )
+    )
+    await Promise.all(
+      objects.versions.map((object) =>
+        deleteObject(bucketName, object.Key, object.VersionId)
+      )
+    )
+
+    logger.info('Successfully emptied S3 bucket', { bucketName })
+  }
 }
 
 const deleteObject = async (
